Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/property-escape-entity-01-expected.txt

CONSOLE MESSAGE: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-property.pl?clutter=blah&q=%22%20onload=%22alert(String.fromCharCode(0x58,0x53,0x53))%26%23x2f%26%2347' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
